Move the blockKeys method to a utility class.
