+++ src/main/java/org/mastodon/revised/trackscheme/display/TrackSchemeSearchPanel.java
@@ -1,6 +1,5 @@
 package org.mastodon.revised.trackscheme.display;
 
-import java.awt.Container;
 import java.awt.Dimension;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
@@ -8,17 +7,12 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.FocusEvent;
 import java.awt.event.FocusListener;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import javax.swing.AbstractAction;
-import javax.swing.Action;
 import javax.swing.Icon;
 import javax.swing.ImageIcon;
-import javax.swing.InputMap;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
 import javax.swing.JComponent;
@@ -40,6 +34,7 @@ import org.mastodon.revised.trackscheme.TrackSchemeFocus;
 import org.mastodon.revised.trackscheme.TrackSchemeGraph;
 import org.mastodon.revised.trackscheme.TrackSchemeNavigation;
 import org.mastodon.revised.trackscheme.TrackSchemeVertex;
+import org.mastodon.revised.ui.util.KeyConfigUtils;
 import org.scijava.ui.behaviour.util.RunnableAction;
 
 import com.itextpdf.text.Font;
@@ -310,7 +305,7 @@ public class TrackSchemeSearchPanel extends JPanel
 			{
 				if ( !keysBlocked )
 				{
-					blockKeys( searchField );
+					KeyConfigUtils.blockKeys( searchField );
 					keysBlocked = true;
 				}
 
@@ -320,56 +315,5 @@ public class TrackSchemeSearchPanel extends JPanel
 				searchField.setForeground( UIManager.getColor( "Label.foreground" ) );
 			}
 		}
-	}
-
-	/**
-	 * TODO move to utility classes.
-	 * <p>
-	 * Preserves the specified component from behavior and action related key
-	 * presses. This useful <i>e.g.</i> for text fields; user input can be
-	 * confused with behaviors key shortcuts.
-	 * <p>
-	 * Adapted from Jan Funke's code in <a href=
-	 * "https://github.com/saalfeldlab/bigcat/blob/janh5/src/main/java/bdv/bigcat/ui/BigCatTable.java#L112-L143">
-	 * BigCat repo</a>
-	 *
-	 * @param ctn
-	 *            the JComponent to preserve from key presses.
-	 */
-	private static void blockKeys( final JComponent ctn )
-	{
-		// Get all keystrokes that are mapped to actions in higher components
-		final ArrayList< KeyStroke > allTableKeys = new ArrayList<>();
-		for ( Container c = ctn.getParent(); c != null; c = c.getParent() )
-		{
-			if ( c instanceof JComponent )
-			{
-				final InputMap inputMap = ( ( JComponent ) c ).getInputMap( JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT );
-				final KeyStroke[] tableKeys = inputMap.allKeys();
-				if ( tableKeys != null )
-					allTableKeys.addAll( Arrays.asList( tableKeys ) );
-			}
-		}
-
-		// An action that does nothing. We can not just map to "none",
-		// as this is not interrupting the action-name -> action search.
-		// We have to map to a proper action, "nothing" in this case.
-		final Action nada = new AbstractAction()
-		{
-			private static final long serialVersionUID = 1L;
-
-			@Override
-			public void actionPerformed( final ActionEvent e )
-			{}
-		};
-		ctn.getActionMap().put( "nothing", nada );
-
-		// Replace every table key binding with nothing, thus creating an
-		// event-barrier.
-		final InputMap inputMap = ctn.getInputMap( JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT );
-		for ( final KeyStroke key : allTableKeys )
-			inputMap.put( key, "nothing" );
-
-		ctn.getActionMap().put( "nothing", nada );
 	}
 }
